AE: Review & reorg.
